+++ Docs/Whats_New.md
@@ -27,7 +27,10 @@ Released TBD
 - Move *Metal* drawable presentation from `MTLCommandBuffer` to `MTLDrawable`
   to improve performance and reduce blocking.
 - Allow binding descriptor set using layout different than it was created with.
+- Report `VkPhysicalDeviceLimits::maxPerStageDescriptorStorageImages` as Metal limit of `8`.
+- Increase per-stage texture count to `96` for A11 SoC's and above.
 - Clarify documentation on mapping limitations for host-coherent image memory on *macOS*.
+- Update `VK_MVK_MOLTENVK_SPEC_VERSION` to `29`.